+++ src/pubsub/tests/emit-self.js
@@ -1,3 +1,4 @@
+// @ts-nocheck interface tests
 /* eslint-env mocha */
 'use strict'
 
@@ -62,7 +63,7 @@ module.exports = (common) => {
         pubsub.publish(topic, data)
 
         // Wait 1 second to guarantee that self is not noticed
-        return new Promise((resolve) => setTimeout(() => resolve(), 1000))
+        return new Promise((resolve) => setTimeout(resolve, 1000))
       })
     })
   })