+++ src/pubsub/tests/messages.js
@@ -1,3 +1,4 @@
+// @ts-nocheck interface tests
 /* eslint-env mocha */
 'use strict'
 
@@ -10,6 +11,7 @@ const uint8ArrayFromString = require('uint8arrays/from-string')
 
 const { utils } = require('..')
 const PeerStreams = require('../peer-streams')
+const { SignaturePolicy } = require('../signature-policy')
 
 const topic = 'foo'
 const data = uint8ArrayFromString('bar')
@@ -31,24 +33,17 @@ module.exports = (common) => {
     })
 
     it('should emit normalized signed messages on publish', async () => {
+      pubsub.globalSignaturePolicy = SignaturePolicy.StrictSign
       sinon.spy(pubsub, '_emitMessage')
-      sinon.spy(utils, 'randomSeqno')
 
       await pubsub.publish(topic, data)
       expect(pubsub._emitMessage.callCount).to.eql(1)
 
       const [messageToEmit] = pubsub._emitMessage.getCall(0).args
 
-      const expected = utils.normalizeInRpcMessage(
-        await pubsub._buildMessage({
-          receivedFrom: pubsub.peerId.toB58String(),
-          from: pubsub.peerId.toB58String(),
-          data,
-          seqno: utils.randomSeqno.getCall(0).returnValue,
-          topicIDs: [topic]
-        }))
-
-      expect(messageToEmit).to.eql(expected)
+      expect(messageToEmit.seqno).to.not.eql(undefined)
+      expect(messageToEmit.key).to.not.eql(undefined)
+      expect(messageToEmit.signature).to.not.eql(undefined)
     })
 
     it('should drop unsigned messages', async () => {
@@ -56,7 +51,10 @@ module.exports = (common) => {
       sinon.spy(pubsub, '_publish')
       sinon.spy(pubsub, 'validate')
 
-      const peerStream = new PeerStreams({ id: await PeerId.create() })
+      const peerStream = new PeerStreams({
+        id: await PeerId.create(),
+        protocol: 'test'
+      })
       const rpc = {
         subscriptions: [],
         msgs: [{
@@ -83,18 +81,20 @@ module.exports = (common) => {
     })
 
     it('should not drop unsigned messages if strict signing is disabled', async () => {
+      pubsub.globalSignaturePolicy = SignaturePolicy.StrictNoSign
       sinon.spy(pubsub, '_emitMessage')
       sinon.spy(pubsub, '_publish')
       sinon.spy(pubsub, 'validate')
-      sinon.stub(pubsub, 'strictSigning').value(false)
 
-      const peerStream = new PeerStreams({ id: await PeerId.create() })
+      const peerStream = new PeerStreams({
+        id: await PeerId.create(),
+        protocol: 'test'
+      })
+
       const rpc = {
         subscriptions: [],
         msgs: [{
-          from: peerStream.id.toBytes(),
           data,
-          seqno: utils.randomSeqno(),
           topicIDs: [topic]
         }]
       }