+++ src/stream-muxer/tests/spawner.js
@@ -1,8 +1,10 @@
+// @ts-nocheck interface tests
 'use strict'
 
 const { expect } = require('chai')
 const pair = require('it-pair/duplex')
-const pipe = require('it-pipe')
+const { pipe } = require('it-pipe')
+
 const pLimit = require('p-limit')
 const { collect, tap, consume } = require('streaming-iterables')
 
@@ -61,8 +63,11 @@ module.exports = async (Muxer, nStreams, nMsg, limit) => {
 }
 
 function marker (n) {
+  /** @type {Function} */
   let check
   let i = 0
+
+  /** @type {Promise<void>} */
   const done = new Promise((resolve, reject) => {
     check = err => {
       i++
@@ -78,5 +83,7 @@ function marker (n) {
       }
     }
   })
+
+  // @ts-ignore - TS can't see that assignement occured
   return { check, done }
 }