+++ src/transport/tests/dial-test.js
@@ -1,3 +1,4 @@
+// @ts-nocheck interface tests
 /* eslint-env mocha */
 'use strict'
 
@@ -9,8 +10,8 @@ chai.use(dirtyChai)
 const { isValidTick } = require('./utils')
 const goodbye = require('it-goodbye')
 const { collect } = require('streaming-iterables')
-const pipe = require('it-pipe')
-const AbortController = require('abort-controller')
+const { pipe } = require('it-pipe')
+const AbortController = require('abort-controller').default
 const AbortError = require('../errors').AbortError
 const sinon = require('sinon')