+++ src/transport/tests/listen-test.js
@@ -1,3 +1,4 @@
+// @ts-nocheck interface tests
 /* eslint max-nested-callbacks: ["error", 8] */
 /* eslint-env mocha */
 'use strict'
@@ -9,7 +10,7 @@ chai.use(dirtyChai)
 const sinon = require('sinon')
 
 const pWaitFor = require('p-wait-for')
-const pipe = require('it-pipe')
+const { pipe } = require('it-pipe')
 const uint8arrayFromString = require('uint8arrays/from-string')
 const { isValidTick } = require('./utils')